split evaluator/partial evaluator into two files
